fix eclipse warnings
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@3770 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ src/java/org/ejbca/core/ejb/services/ServiceTimerSessionBean.java
@@ -174,7 +174,7 @@ public class ServiceTimerSessionBean extends BaseSessionBean implements javax.ej
 		try{
 		  UserTransaction ut = getSessionContext().getUserTransaction();
 		  ut.begin();
-	      serviceData = getServiceSession().getServiceConfiguration(intAdmin, timerInfo);
+	      serviceData = getServiceSession().getServiceConfiguration(intAdmin, timerInfo.intValue());
 	      if(serviceData != null){
 	    	  serviceName = getServiceSession().getServiceName(intAdmin, timerInfo.intValue());
 		      worker = getWorker(serviceData,serviceName);
